refactor : 여행 상품 관련 컴포넌트 폴더로 묶음
컴포넌트가 많아져서 같은 페이지에서 사용되는 컴포넌트 폴더로 관리
